+++ changelog.md
@@ -8,6 +8,7 @@
 - Added check that a task cannot be planned on a day, when the time it would be planned is less than the min_slot_length setting, but the task would normally have more remaining than that. In that case, technically, a task will be planned on the other days instead.
 - Min slot length won't be checked for running tasks now.
 - Changed that still open tasks without remaining time won't use the score * non_time_mode_minutes anymore as the planned remaining time. Instead is is 1 minute now.
+- Changed that "clean done tasks" function will set subtasks with status "1" to status "0" now.
 
 
 ## v3.1