+++ changelog.md
@@ -9,6 +9,7 @@
 - Min slot length won't be checked for running tasks now.
 - Changed that still open tasks without remaining time won't use the score * non_time_mode_minutes anymore as the planned remaining time. Instead is is 1 minute now.
 - Changed that "clean done tasks" function will set subtasks with status "1" to status "0" now.
+- Added that for the tasks in the automatic plan there is now the "planned" time together with the "total reamining" time on the text output and the automatic plan page (+ sticky).
 
 
 ## v3.1